requirements.txt
aiogram
transformers
torch
python-dotenv
